requirements.txt
datasets
huggingface_hub
openai
pillow
tqdm
requests
evaluate
sacrebleu
torch
gradio
pandas
tenacity
beautifulsoup4